Extract BM25 logic into dedicated struct and split text index into files
